fix: correction of the body modification
A custom type was created and configured to recognize the uid field of the body that has the  request. This with the objective of being able to add this field to the body and to be able to access it in the different controllers that are handled and also so that typescript does not complain.